+++ eng/common/scripts/ChangeLog-Operations.ps1
@@ -139,7 +139,8 @@ function Confirm-ChangeLogEntry {
     [String]$VersionString,
     [boolean]$ForRelease = $false,
     [Switch]$SantizeEntry,
-    [PSCustomObject]$ChangeLogStatus = $null
+    [PSCustomObject]$ChangeLogStatus = $null,
+    [boolean]$SuppressErrors = $false
   )
 
   if (!$ChangeLogStatus) {
@@ -148,17 +149,15 @@ function Confirm-ChangeLogEntry {
       Message = ""
     }
   }
-  else {
-    # Do not stop the script on error when status object is passed as param
-    $ErrorActionPreference = 'Continue'
-  }
   $changeLogEntries = Get-ChangeLogEntries -ChangeLogLocation $ChangeLogLocation
   $changeLogEntry = $changeLogEntries[$VersionString]
 
   if (!$changeLogEntry) {
     $ChangeLogStatus.Message = "ChangeLog[${ChangeLogLocation}] does not have an entry for version ${VersionString}."
     $ChangeLogStatus.IsValid = $false
-    LogError "$($ChangeLogStatus.Message)"
+    if (!$SuppressErrors) {
+      LogError "$($ChangeLogStatus.Message)"
+    }
     return $false
   }
 
@@ -176,14 +175,16 @@ function Confirm-ChangeLogEntry {
   if ([System.String]::IsNullOrEmpty($changeLogEntry.ReleaseStatus)) {
     $ChangeLogStatus.Message = "Entry does not have a release status. Please ensure the status is set to a date '($CHANGELOG_DATE_FORMAT)' or '$CHANGELOG_UNRELEASED_STATUS' if not yet released. See https://aka.ms/azsdk/guideline/changelogs for more info."
     $ChangeLogStatus.IsValid = $false
-    LogError "$($ChangeLogStatus.Message)"
+    if (!$SuppressErrors) {
+      LogError "$($ChangeLogStatus.Message)"
+    }
     return $false
   }
 
   if ($ForRelease -eq $True)
   {
     LogDebug "Verifying as a release build because ForRelease parameter is set to true"
-    return Confirm-ChangeLogForRelease -changeLogEntry $changeLogEntry -changeLogEntries $changeLogEntries -ChangeLogStatus $ChangeLogStatus
+    return Confirm-ChangeLogForRelease -changeLogEntry $changeLogEntry -changeLogEntries $changeLogEntries -ChangeLogStatus $ChangeLogStatus -SuppressErrors $SuppressErrors
   }
 
   # If the release status is a valid date then verify like its about to be released
@@ -191,7 +192,7 @@ function Confirm-ChangeLogEntry {
   if ($status -as [DateTime])
   {
     LogDebug "Verifying as a release build because the changelog entry has a valid date."
-    return Confirm-ChangeLogForRelease -changeLogEntry $changeLogEntry -changeLogEntries $changeLogEntries -ChangeLogStatus $ChangeLogStatus
+    return Confirm-ChangeLogForRelease -changeLogEntry $changeLogEntry -changeLogEntries $changeLogEntries -ChangeLogStatus $ChangeLogStatus -SuppressErrors $SuppressErrors
   }
 
   $ChangeLogStatus.Message = "ChangeLog[${ChangeLogLocation}] has an entry for version ${VersionString}."
@@ -356,7 +357,8 @@ function Confirm-ChangeLogForRelease {
     $changeLogEntry,
     [Parameter(Mandatory = $true)]
     $changeLogEntries,
-    $ChangeLogStatus = $null
+    $ChangeLogStatus = $null,
+    $SuppressErrors = $false
   )
 
   if (!$ChangeLogStatus) {
@@ -365,13 +367,16 @@ function Confirm-ChangeLogForRelease {
       Message = ""
     }
   }
+
   $entries = Sort-ChangeLogEntries -changeLogEntries $changeLogEntries
 
   $ChangeLogStatus.IsValid = $true
   if ($changeLogEntry.ReleaseStatus -eq $CHANGELOG_UNRELEASED_STATUS) {
     $ChangeLogStatus.Message = "Entry has no release date set. Please ensure to set a release date with format '$CHANGELOG_DATE_FORMAT'. See https://aka.ms/azsdk/guideline/changelogs for more info."
     $ChangeLogStatus.IsValid = $false
-    LogError "$($ChangeLogStatus.Message)"
+    if (!$SuppressErrors) {
+      LogError "$($ChangeLogStatus.Message)"
+    }
   }
   else {
     $status = $changeLogEntry.ReleaseStatus.Trim().Trim("()")
@@ -381,27 +386,35 @@ function Confirm-ChangeLogForRelease {
       {
         $ChangeLogStatus.Message = "Date must be in the format $($CHANGELOG_DATE_FORMAT). See https://aka.ms/azsdk/guideline/changelogs for more info."
         $ChangeLogStatus.IsValid = $false
-        LogError "$($ChangeLogStatus.Message)"
+        if (!$SuppressErrors) {
+          LogError "$($ChangeLogStatus.Message)"
+        }
       }
 
       if (@($entries.ReleaseStatus)[0] -ne $changeLogEntry.ReleaseStatus)
       {
         $ChangeLogStatus.Message = "Invalid date [ $status ]. The date for the changelog being released must be the latest in the file."
         $ChangeLogStatus.IsValid = $false
-        LogError "$($ChangeLogStatus.Message)"
+        if (!$SuppressErrors) {
+          LogError "$($ChangeLogStatus.Message)"
+        }
       }
     }
     catch {
         $ChangeLogStatus.Message = "Invalid date [ $status ] passed as status for Version [$($changeLogEntry.ReleaseVersion)]. See https://aka.ms/azsdk/guideline/changelogs for more info."
         $ChangeLogStatus.IsValid = $false
-        LogError "$($ChangeLogStatus.Message)"
+        if (!$SuppressErrors) {
+          LogError "$($ChangeLogStatus.Message)"
+        }
     }
   }
 
   if ([System.String]::IsNullOrWhiteSpace($changeLogEntry.ReleaseContent)) {
     $ChangeLogStatus.Message = "Entry has no content. Please ensure to provide some content of what changed in this version. See https://aka.ms/azsdk/guideline/changelogs for more info."
     $ChangeLogStatus.IsValid = $false
-    LogError "$($ChangeLogStatus.Message)"
+    if (!$SuppressErrors) {
+      LogError "$($ChangeLogStatus.Message)"
+    }
   }
 
   $foundRecommendedSection = $false
@@ -422,12 +435,16 @@ function Confirm-ChangeLogForRelease {
   {
     $ChangeLogStatus.Message = "The changelog entry has the following sections with no content ($($emptySections -join ', ')). Please ensure to either remove the empty sections or add content to the section."
     $ChangeLogStatus.IsValid = $false
-    LogError "$($ChangeLogStatus.Message)"
+    if (!$SuppressErrors) {
+      LogError "$($ChangeLogStatus.Message)"
+    }
   }
   if (!$foundRecommendedSection)
   {
     $ChangeLogStatus.Message = "The changelog entry did not contain any of the recommended sections ($($RecommendedSectionHeaders -join ', ')), please add at least one. See https://aka.ms/azsdk/guideline/changelogs for more info."
-    LogWarning "$($ChangeLogStatus.Message)"
+    if (!$SuppressErrors) {
+      LogError "$($ChangeLogStatus.Message)"
+    }
   }
   return $ChangeLogStatus.IsValid
 }