+++ sdk/botservice/Azure.ResourceManager.BotService/tests/TestCase/ConnectionSettingTest.cs
@@ -22,6 +22,7 @@ namespace Azure.ResourceManager.BotService.Tests
 
         [TestCase]
         [RecordedTest]
+        [LiveOnly(Reason = "https://github.com/Azure/azure-sdk-for-net/issues/43404")]
         public async Task CreateOrUpdate()
         {
             //1.Create
@@ -80,6 +81,7 @@ namespace Azure.ResourceManager.BotService.Tests
 
         [TestCase]
         [RecordedTest]
+        [LiveOnly(Reason = "https://github.com/Azure/azure-sdk-for-net/issues/43404")]
         public async Task ListTest()
         {
             //Create