+++ sdk/eventhub/Microsoft.Azure.WebJobs.Extensions.EventHubs/tests/EventHubConfigurationTests.cs
@@ -44,6 +44,13 @@ namespace Microsoft.Azure.WebJobs.EventHubs.UnitTests
         }
 
         [Test]
+        public void ConfigureOptions_CheckpointingEnabledByDefault()
+        {
+            EventHubOptions options = CreateOptionsFromConfigWithoutCheckpointEnabled();
+            Assert.True(options.EnableCheckpointing);
+        }
+
+        [Test]
         public void ConfigureOptions_AppliesValuesCorrectly()
         {
             EventHubOptions options = CreateOptionsFromConfig();
@@ -66,6 +73,7 @@ namespace Microsoft.Azure.WebJobs.EventHubs.UnitTests
             Assert.AreEqual(EventHubsTransportType.AmqpWebSockets, options.TransportType);
             Assert.AreEqual("http://proxyserver:8080/", ((WebProxy) options.WebProxy).Address.AbsoluteUri);
             Assert.AreEqual("http://www.customendpoint.com/", options.CustomEndpointAddress.ToString());
+            Assert.False(options.EnableCheckpointing);
         }
 
         [Test]
@@ -99,6 +107,7 @@ namespace Microsoft.Azure.WebJobs.EventHubs.UnitTests
             Assert.AreEqual(EventHubsTransportType.AmqpWebSockets, result.TransportType);
             Assert.AreEqual("http://proxyserver:8080/", ((WebProxy) result.WebProxy).Address.AbsoluteUri);
             Assert.AreEqual("http://www.customendpoint.com/", result.CustomEndpointAddress.AbsoluteUri);
+            Assert.False(result.EnableCheckpointing);
         }
 
         [Test]
@@ -115,6 +124,7 @@ namespace Microsoft.Azure.WebJobs.EventHubs.UnitTests
             Assert.AreEqual(21, options.EventProcessorOptions.LoadBalancingUpdateInterval.TotalSeconds);
             Assert.AreEqual("FromEnqueuedTime", options.InitialOffsetOptions.Type.ToString());
             Assert.AreEqual("2020-09-13 12:00:00Z", options.InitialOffsetOptions.EnqueuedTimeUtc.Value.ToString("u"));
+            Assert.IsTrue(options.EnableCheckpointing);
         }
 
         [Test]
@@ -140,6 +150,7 @@ namespace Microsoft.Azure.WebJobs.EventHubs.UnitTests
             Assert.AreEqual(TimeSpan.FromSeconds(21), result.LoadBalancingUpdateInterval);
             Assert.AreEqual("FromEnqueuedTime", result.InitialOffsetOptions.Type.ToString());
             Assert.AreEqual("2020-09-13 12:00:00Z", result.InitialOffsetOptions.EnqueuedTimeUtc.Value.ToString("u"));
+            Assert.IsTrue(options.EnableCheckpointing);
         }
 
         [Test]
@@ -272,6 +283,7 @@ namespace Microsoft.Azure.WebJobs.EventHubs.UnitTests
                 { $"{extensionPath}:TransportType", "amqpWebSockets" },
                 { $"{extensionPath}:WebProxy", "http://proxyserver:8080/" },
                 { $"{extensionPath}:CustomEndpointAddress", "http://www.customendpoint.com/" },
+                { $"{extensionPath}:{nameof(EventHubOptions.EnableCheckpointing)}", "false" },
             };
 
             return TestHelpers.GetConfiguredOptions<EventHubOptions>(b =>
@@ -296,6 +308,38 @@ namespace Microsoft.Azure.WebJobs.EventHubs.UnitTests
                 { $"{extensionPath}:PartitionManagerOptions:RenewInterval", "00:00:21" },
                 { $"{extensionPath}:InitialOffsetOptions:Type", "FromEnqueuedTime" },
                 { $"{extensionPath}:InitialOffsetOptions:EnqueuedTimeUTC", "2020-09-13 12:00:00Z" },
+            };
+
+            return TestHelpers.GetConfiguredOptions<EventHubOptions>(b =>
+            {
+                b.AddEventHubs();
+            }, values);
+        }
+
+        private EventHubOptions CreateOptionsFromConfigWithoutCheckpointEnabled()
+        {
+            string extensionPath = "AzureWebJobs:Extensions:EventHubs";
+            var values = new Dictionary<string, string>
+            {
+                { $"{extensionPath}:MaxEventBatchSize", "123" },
+                { $"{extensionPath}:MinEventBatchSize", "100" },
+                { $"{extensionPath}:MaxWaitTime", "00:01:00" },
+                { $"{extensionPath}:TrackLastEnqueuedEventProperties", "true" },
+                { $"{extensionPath}:PrefetchCount", "123" },
+                { $"{extensionPath}:BatchCheckpointFrequency", "5" },
+                { $"{extensionPath}:PartitionOwnershipExpirationInterval", "00:00:31" },
+                { $"{extensionPath}:LoadBalancingUpdateInterval", "00:00:21" },
+                { $"{extensionPath}:LoadBalancingStrategy", "greedy" },
+                { $"{extensionPath}:InitialOffsetOptions:Type", "FromEnqueuedTime" },
+                { $"{extensionPath}:InitialOffsetOptions:EnqueuedTimeUTC", "2020-09-13 12:00:00Z" },
+                { $"{extensionPath}:ClientRetryOptions:MaximumRetries", "5" },
+                { $"{extensionPath}:ClientRetryOptions:Delay", "00:00:01" },
+                { $"{extensionPath}:ClientRetryOptions:MaxDelay", "00:01:00" },
+                { $"{extensionPath}:ClientRetryOptions:TryTimeout", "00:01:30" },
+                { $"{extensionPath}:ClientRetryOptions:Mode", "0" },
+                { $"{extensionPath}:TransportType", "amqpWebSockets" },
+                { $"{extensionPath}:WebProxy", "http://proxyserver:8080/" },
+                { $"{extensionPath}:CustomEndpointAddress", "http://www.customendpoint.com/" },
             };
 
             return TestHelpers.GetConfiguredOptions<EventHubOptions>(b =>