+++ sdk/sqlmanagement/Azure.ResourceManager.Sql/CHANGELOG.md
@@ -1,6 +1,6 @@
 # Release History
 
-## 1.3.0-beta.7 (Unreleased)
+## 1.3.0-beta.8 (Unreleased)
 
 ### Features Added
 
@@ -9,6 +9,12 @@
 ### Bugs Fixed
 
 ### Other Changes
+
+## 1.3.0-beta.7 (2024-04-29)
+
+### Features Added
+
+- Add `ArmOperation.Rehydrate` and `ArmOperation.Rehydrate<T>` static methods to rehydrate a long-running operation.
 
 ## 1.3.0-beta.6 (2024-03-26)